Extract PDF import persistence
