+++ service-worker.js
@@ -116,6 +116,7 @@ const APP_SHELL = [
   '/js/pdf-import-review.js',
   '/js/pdf-import-marker-mapping.js',
   '/js/pdf-import-marker-normalization.js',
+  '/js/pdf-import-persistence.js',
   '/js/export.js',
   '/js/chat.js',
   '/js/chat-window-bindings.js',
