+++ tests/test-audit.js
@@ -62,7 +62,8 @@ assert('SW APP_SHELL includes PDF import review module', swAuditSrc.includes("'/
 assert('SW APP_SHELL includes PDF import support modules',
   swAuditSrc.includes("'/js/pdf-import-preflight.js'")
   && swAuditSrc.includes("'/js/pdf-import-progress.js'")
-  && swAuditSrc.includes("'/js/pdf-import-marker-normalization.js'"));
+  && swAuditSrc.includes("'/js/pdf-import-marker-normalization.js'")
+  && swAuditSrc.includes("'/js/pdf-import-persistence.js'"));
 assert('SW APP_SHELL includes context card summary module', swAuditSrc.includes("'/js/context-card-summaries.js'"));
 assert('SW APP_SHELL includes context card editor UI module', swAuditSrc.includes("'/js/context-card-editor-ui.js'"));
 assert('SW APP_SHELL includes context card medical history module', swAuditSrc.includes("'/js/context-card-medical-history-editor.js'"));
