+++ tests/test-correctness-phase2.js
@@ -103,6 +103,7 @@ const pwaAppShellAssets = [
   '/js/pdf-import-review.js',
   '/js/pdf-import-marker-mapping.js',
   '/js/pdf-import-marker-normalization.js',
+  '/js/pdf-import-persistence.js',
   '/js/blob-storage.js',
   '/js/data-merge.js',
   '/js/views-router.js',
